pyproject.toml
[tool.poetry]
name = "agentSerpWorker"
version = "0.1.0"
description = "a serp worker build with langgraph and notion api"
authors = ["Frank Lin <etrobot@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langgraph = "^0.2.14"
python-dotenv = "^1.0.1"
langchain-openai = "^0.1.23"
notion-client = "^2.2.1"
langgraph-checkpoint-sqlite = "^1.0.3"
shortuuid = "^1.0.13"
aiosqlite = "^0.20.0"
langchain-community = "^0.2.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
